# docs/swagger.yaml
components:
  schemas:
    handlers.createRecitationDTO:
      properties:
        slug:
          type: string
      type: object
...
    handlers.updateRecitationDTO:
      properties:
        slug:
          type: string
      type: object
...
    models.Error:
      properties:
        error:
          type: string
        message:
          type: string
      type: object
...
    sqlc.Recitation:
      properties:
        name:
          type: string
        reciter:
          type: string
        slug:
          type: string
      type: object
    sqlc.RecitationFile:
      properties:
        has_timings:
          type: boolean
        lafzize_processing:
          type: boolean
        reciter:
          type: string
        slug:
          type: string
        verse_key:
          type: string
      type: object
...
info:
  title: tilawah-hub
  version: 0.0.1
openapi: 3.1.0
paths:
...
  /recitation-files/{reciter}/{slug}:
    get:
      parameters:
      - description: Reciter
        in: path
        name: reciter
        required: true
        schema:
          type: string
      - description: Slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/sqlc.RecitationFile'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - RecitationFile
  /recitation-files/{reciter}/{slug}/{verse_key}:
    get:
      parameters:
      - description: Reciter
        in: path
        name: reciter
        required: true
        schema:
          type: string
      - description: Slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Verse key
        in: path
        name: verse_key
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.RecitationFile'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - RecitationFile
  /recitation-files/{slug}/:
    post:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      - description: Slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: string
        description: Verse Key
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.RecitationFile'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - RecitationFile
  /recitation-files/{slug}/{verse_key}:
    delete:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      - description: Slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Verse Key
        in: path
        name: verse_key
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.RecitationFile'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
      tags:
      - RecitationFile
...
  /recitations:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/sqlc.Recitation'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - Recitation
    post:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.createRecitationDTO'
        description: Create Recitation
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.Recitation'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - Recitation
  /recitations/{reciter}/{slug}:
    get:
      parameters:
      - description: Reciter
        in: path
        name: reciter
        required: true
        schema:
          type: string
      - description: Slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.Recitation'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - Recitation
  /recitations/{slug}:
    delete:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      - description: Slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.Recitation'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - Recitation
    put:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      - description: Slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.updateRecitationDTO'
        description: Update Recitation
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.Recitation'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - Recitation
...